+++ NEWS.md
@@ -1,3 +1,10 @@
+# y2artisan 0.5.9
+## New functions
+* ms_scatter_y2 - Creates an mschart object automatically formatted for a scatter plot. This function requires two lists called "text_settings" and "color_settings" and a named vector called "size_settings".
+## New arguments
+* gg_dotplot_y2 - added alpha argument.
+
+
 # y2artisan 0.5.8
 ## New functions
 * gg_dotplot_y2 - Creates a ggplot2 object automatically formatted for a dot plot with a grouping variable.